chest setting

+++ src/main/kotlin/com/lambda/utils/Timer.kt
@@ -3,7 +3,7 @@ package com.lambda.utils
 class Timer {
     private var time: Long = -1
 
-    fun passed(ms: Double): Boolean {
+    fun passed(ms: Long): Boolean {
         return System.currentTimeMillis() - time >= ms
     }
